Separate MessageGroup and MessageItem components
