+++ frontend/app/chat/page.tsx
@@ -1,64 +1,14 @@
 "use client";
-import { HStack, Stack } from "@/app/ui/layout/stack";
+import { Stack } from "@/app/ui/layout/stack";
 import { Input } from "@/components/ui/input";
 import { Button } from "@/components/ui/button";
 import { useEffect, useRef, useState } from "react";
 import { Separator } from "@/components/ui/separator";
 import { users, messages } from "./test-data";
 import type { Message } from "./test-data";
-import { AvatarSkeleton, MessageSkeleton } from "./skeleton";
+import { MessageSkeleton } from "./skeleton";
 import { Sidebar } from "./sidebar";
-
-function MessageItem({
-  message,
-  withAvatar,
-}: {
-  message: Message;
-  withAvatar: boolean;
-}) {
-  const created_at_hhmm = message.created_at.split(" ")[1].slice(0, 5);
-
-  return (
-    <HStack spacing={4} className="group hover:opacity-60">
-      {withAvatar && (
-        <>
-          {/* Left Side */}
-          <AvatarSkeleton />
-          {/* Right Side */}
-          <Stack>
-            <HStack spacing={2}>
-              <div className="text-xs">{message.user_name}</div>
-              <div className="text-xs text-muted-foreground">
-                {message.created_at}
-              </div>
-            </HStack>
-            <div className="text-sm text-muted-foreground">{message.text}</div>
-          </Stack>
-        </>
-      )}
-      {!withAvatar && (
-        <>
-          {/* Left Side */}
-          <div className="group-hover:text-muted-foreground flex-none text-background text-xs w-10 text-center">
-            {created_at_hhmm}
-          </div>
-          {/* Right Side */}
-          <div className="text-sm text-muted-foreground">{message.text}</div>
-        </>
-      )}
-    </HStack>
-  );
-}
-
-function MessageGroup({ messages }: { messages: Message[] }) {
-  return (
-    <Stack spacing={1}>
-      {messages.map((msg, i) => {
-        return <MessageItem message={msg} withAvatar={i === 0} key={msg.id} />;
-      })}
-    </Stack>
-  );
-}
+import { MessageGroup } from "./message-group";
 
 function groupMessagesByUser(messages: Message[]): Message[][] {
   let prevUserId: number | undefined = undefined;
